Feat.Status-PC

See merge request s07-final/S07P31B309!82

+++ ec2-back/gganbu/src/main/java/com/ssafy/gganbu/model/SocketVO.java
@@ -15,7 +15,7 @@ public class SocketVO {
     // 메시지의 내용을 저장하기 위한 변
     private String task;
 
-    private Long status;
+    private int status;
 
     public SocketVO() {
